+++ src/blocks/block.rs
@@ -8,7 +8,9 @@ use ratatui::{
 // ============================================================================
 // Block Variants (Gems)
 // ============================================================================
-// TODO Change the colors. Rename "Gems"? Verfiy that all colors are "standard" (e.g. Rgb(255, 165, 0)).
+// TODO Change the colors
+// TODO Rename `Gem`
+// FIXME Use `Color::Red` style or `Color::Indexed(0..255)` throughout the app
 const STYLE_RUBY: Style = Style::new().bg(Color::Red);
 const STYLE_AMBER: Style = Style::new().bg(Color::Rgb(255, 165, 0));
 const STYLE_TOPAZ: Style = Style::new().bg(Color::Yellow);
@@ -55,9 +57,9 @@ define_block_variants!(
 // Block
 // ============================================================================
 pub struct Block {
-    pub x: u8,
-    pub y: i8,
-    pub variant: Gem,
+    x: u8,
+    y: i8,
+    variant: Gem,
 }
 
 impl Block {