+++ src/blocks/column.rs
@@ -3,66 +3,53 @@ use ratatui::{buffer::Buffer, layout::Rect, widgets::Widget};
 use crate::blocks::{Block, Gem};
 
 pub struct Column {
-    pub blocks: [Block; 3],
+    x: u8,
+    y: i8,
+    gems: [Gem; 3],
 }
 
 impl Column {
     pub fn new(x: u8, y: i8, rng: &mut fastrand::Rng) -> Self {
-        Self {
-            blocks: [
-                Block::new(x, y, Gem::random(rng)),
-                Block::new(x, y + 1, Gem::random(rng)),
-                Block::new(x, y + 2, Gem::random(rng)),
-            ],
-        }
+        Self { x, y, gems: [Gem::random(rng), Gem::random(rng), Gem::random(rng)] }
     }
 
-    pub fn move_down(&mut self) {
-        for block in &mut self.blocks {
-            block.y += 1;
-        }
+    pub const fn view_gems(&self) -> [(u8, i8, Gem); 3] {
+        [(self.x, self.y, self.gems[0]), (self.x, self.y + 1, self.gems[1]), (self.x, self.y + 2, self.gems[2])]
     }
 
-    pub fn move_left(&mut self) {
-        for block in &mut self.blocks {
-            if block.x > 0 {
-                block.x -= 1;
-            }
-        }
+    pub const fn x(&self) -> u8 {
+        self.x
     }
 
-    pub fn move_right(&mut self) {
-        for block in &mut self.blocks {
-            block.x += 1;
-        }
+    pub const fn y_bottom(&self) -> i8 {
+        self.y + 2
+    }
+
+    pub const fn move_down(&mut self, distance: i8) {
+        self.y += distance;
+    }
+
+    pub const fn move_left(&mut self) {
+        self.x -= 1;
+    }
+
+    pub const fn move_right(&mut self) {
+        self.x += 1;
     }
 
     pub const fn rotate_up(&mut self) {
-        let temp = self.blocks[0].variant;
-        self.blocks[0].variant = self.blocks[1].variant;
-        self.blocks[1].variant = self.blocks[2].variant;
-        self.blocks[2].variant = temp;
+        self.gems.rotate_left(1);
     }
 
     pub const fn rotate_down(&mut self) {
-        let temp = self.blocks[2].variant;
-        self.blocks[2].variant = self.blocks[1].variant;
-        self.blocks[1].variant = self.blocks[0].variant;
-        self.blocks[0].variant = temp;
-    }
-
-    pub fn drop(&mut self, distance: i8) {
-        if distance > 0 {
-            for block in &mut self.blocks {
-                block.y += distance;
-            }
-        }
+        self.gems.rotate_right(1);
     }
 }
 
 impl Widget for &Column {
     fn render(self, area: Rect, buf: &mut Buffer) {
-        for block in &self.blocks {
+        for (i, gem) in self.gems.iter().enumerate() {
+            let block = Block::new(self.x, self.y + i as i8, *gem);
             block.render(area, buf);
         }
     }