+++ src/blocks/mod.rs
@@ -1,5 +1,7 @@
 pub mod block;
 pub mod column;
+pub mod pile;
 
 pub use block::{Block, Gem};
 pub use column::Column;
+pub use pile::Pile;